requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
jinja2
aiofiles
